# public/swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  Account:
    description: A tenant account
    example:
      created_at: 2008-01-22T00:07:16-08:00
      created_by: Vero est mollitia eos voluptas.
      href: Enim suscipit est explicabo eveniet dolore.
      id: 7.9748073e+18
      name: Saepe eum cumque sint veniam.
    properties:
      created_at:
        description: Date of creation
        example: 2015-09-23T11:51:57-07:00
        format: date-time
        type: string
      created_by:
        description: Email of account owner
        example: loyce.doyle@pollich.com
        format: email
        type: string
      href:
        description: API href of account
        example: Et ut omnis praesentium non.
        type: string
      id:
        description: ID of account
        example: 6.955004e+18
        format: int64
        type: integer
      name:
        description: Name of account
        example: Explicabo laboriosam consequatur.
        type: string
    required:
    - id
    - href
    - name
    - created_at
    - created_by
    title: 'Mediatype identifier: application/vnd.account+json'
    type: object
  AccountCollection:
    example:
    - created_at: 1978-02-18T21:20:57-08:00
      created_by: Reiciendis ea architecto magni tempora.
      href: Distinctio delectus vel earum doloribus.
      id: 6.5796815e+18
      name: Recusandae molestiae nemo omnis quae suscipit.
    - created_at: 1972-08-12T17:28:45-07:00
      created_by: Sapiente et sunt dolorem.
      href: Ea est.
      id: 3.228513e+17
      name: Sequi blanditiis dolore laudantium repellat perspiciatis iure.
    - created_at: 2006-10-10T05:51:30-07:00
      created_by: Et ut repellat voluptatem voluptatem voluptas veritatis.
      href: Dignissimos sunt ut accusamus soluta.
      id: 1.3258674e+18
      name: Dolore et.
    items:
      $ref: '#/definitions/Account'
    title: 'Mediatype identifier: application/vnd.account+json; type=collection'
    type: array
  Bottle:
    description: A bottle of wine
    example:
      account:
        created_at: 2012-05-30T13:43:59-07:00
        created_by: Repellendus quidem modi iste incidunt.
        href: Omnis eveniet voluptatem.
        id: 5.3925295e+18
        name: Et sunt sit dolor magnam dolores.
      color: Quibusdam aut quibusdam.
      country: Ex quis aut rem.
      created_at: 2002-11-08T19:24:24-08:00
      href: Aut animi nulla aut aut.
      id: 5.128319e+18
      name: Aliquid facere.
      rating: 8.590149e+18
      region: Ea vel.
      review: Incidunt earum dolorem suscipit ut iusto.
      sweetness: 6.5962424e+18
      updated_at: 1995-11-13T20:26:35-08:00
      varietal: Harum qui natus et aut.
      vineyard: Minus rerum ut libero tempore optio.
      vintage: 8.348819e+18
    properties:
      account:
        $ref: '#/definitions/Account'
      color:
        enum:
        - red
        - white
        - rose
        - yellow
        - sparkling
        example: red
        type: string
      country:
        example: wjin
        minLength: 2
        type: string
      created_at:
        description: Date of creation
        example: 2011-12-29T17:17:22-08:00
        format: date-time
        type: string
      href:
        description: API href of bottle
        example: Rem voluptatem.
        type: string
      id:
        description: ID of bottle
        example: 5.5720336e+18
        format: int64
        type: integer
      name:
        example: 0v
        minLength: 2
        type: string
      rating:
        description: Rating of bottle between 1 and 5
        example: 4
        maximum: 5
        minimum: 1
        type: integer
      region:
        example: Enim nam nemo pariatur quam quo laudantium.
        type: string
      review:
        example: gx4
        maxLength: 300
        minLength: 3
        type: string
      sweetness:
        example: 2
        maximum: 5
        minimum: 1
        type: integer
      updated_at:
        description: Date of last update
        example: 2011-07-06T02:59:22-07:00
        format: date-time
        type: string
      varietal:
        example: ulxw
        minLength: 4
        type: string
      vineyard:
        example: w8
        minLength: 2
        type: string
      vintage:
        example: 1958
        maximum: 2020
        minimum: 1900
        type: integer
    required:
    - id
    - href
    - name
    - vineyard
    - varietal
    - vintage
    - color
    - created_at
    title: 'Mediatype identifier: application/vnd.bottle+json'
    type: object
  BottleCollection:
    example:
    - account:
        created_at: 1981-07-23T18:31:00-07:00
        created_by: Odit aperiam voluptatem aut.
        href: Vel molestiae culpa incidunt omnis nobis nobis.
        id: 2.2154884e+18
        name: Totam minus nihil iure.
      color: Est fugiat.
      country: Occaecati quas doloremque et laborum fugit eos.
      created_at: 2013-12-12T09:48:24-08:00
      href: Dolore iusto perspiciatis nobis assumenda voluptas voluptas.
      id: 7.369533e+18
      name: Aperiam quam.
      rating: 4.0068974e+18
      region: Sit et.
      review: In est.
      sweetness: 3.3761357e+17
      updated_at: 2001-05-22T00:40:52-07:00
      varietal: Et qui odio dolore atque.
      vineyard: Ea praesentium repudiandae quia.
      vintage: 9.943361e+17
    - account:
        created_at: 1980-03-06T03:07:52-08:00
        created_by: Dicta impedit est quia corrupti.
        href: Vero facere harum dolores adipisci.
        id: 8.1852995e+18
        name: Esse quia ratione.
      color: Quae deleniti fugit.
      country: Dolor est consequuntur sed.
      created_at: 2006-10-04T13:31:05-07:00
      href: Voluptatem omnis magni.
      id: 2.5622392e+18
      name: Praesentium cum quaerat accusamus voluptatem et fugit.
      rating: 1.836711e+18
      region: Consequatur ut explicabo libero qui aut.
      review: Deleniti et beatae nisi soluta ex.
      sweetness: 4.1388346e+17
      updated_at: 2014-07-23T19:01:04-07:00
      varietal: Rerum dicta quo.
      vineyard: Optio aspernatur modi.
      vintage: 4.5090477e+18
    - account:
        created_at: 1970-06-04T11:54:24-07:00
        created_by: Et et iusto consequatur recusandae.
        href: Vitae libero quis non soluta.
        id: 5.6490115e+18
        name: Nobis quia dolorum eius quo fugit ab.
      color: Ad cumque.
      country: Dolorum eos dolore ipsum eius voluptatem.
      created_at: 1986-08-26T02:32:51-07:00
      href: Laudantium corrupti et.
      id: 6.012384e+18
      name: Modi sed et nulla.
      rating: 3.6717306e+18
      region: Aperiam amet eum officiis eligendi.
      review: Facere quaerat dicta corporis.
      sweetness: 5.517688e+17
      updated_at: 2008-04-22T03:12:28-07:00
      varietal: Eaque repellat quidem laudantium eaque dignissimos qui.
      vineyard: Sit aut officia pariatur.
      vintage: 1.1044081e+18
    items:
      $ref: '#/definitions/Bottle'
    title: 'Mediatype identifier: application/vnd.bottle+json; type=collection'
    type: array
  CreateAccountPayload:
    properties:
      name:
        description: Name of account
        type: string
    required:
    - name
    title: CreateAccountPayload
    type: object
  CreateBottlePayload:
    properties:
      color:
        enum:
        - red
        - white
        - rose
        - yellow
        - sparkling
        type: string
      country:
        minLength: 2
        type: string
      name:
        minLength: 2
        type: string
      region:
        type: string
      review:
        maxLength: 300
        minLength: 3
        type: string
      sweetness:
        maximum: 5
        minimum: 1
        type: integer
      varietal:
        minLength: 4
        type: string
      vineyard:
        minLength: 2
        type: string
      vintage:
        maximum: 2020
        minimum: 1900
        type: integer
    required:
    - name
    - vineyard
    - varietal
    - vintage
    - color
    title: CreateBottlePayload
    type: object
...
  RateBottlePayload:
    properties:
      rating:
        description: Rating of bottle between 1 and 5
        maximum: 5
        minimum: 1
        type: integer
    required:
    - rating
    title: RateBottlePayload
    type: object
  UpdateAccountPayload:
    properties:
      name:
        description: Name of account
        type: string
    required:
    - name
    title: UpdateAccountPayload
    type: object
  UpdateBottlePayload:
    properties:
      color:
        enum:
        - red
        - white
        - rose
        - yellow
        - sparkling
        type: string
      country:
        minLength: 2
        type: string
      name:
        minLength: 2
        type: string
      region:
        type: string
      review:
        maxLength: 300
        minLength: 3
        type: string
      sweetness:
        maximum: 5
        minimum: 1
        type: integer
      varietal:
        minLength: 4
        type: string
      vineyard:
        minLength: 2
        type: string
      vintage:
        maximum: 2020
        minimum: 1900
        type: integer
    title: UpdateBottlePayload
    type: object
...
host: localhost:8081
info:
  contact:
    email: admin@goa.design
    name: goa team
    url: http://goa.design
  description: A basic example of a CRUD API implemented with goa
  license:
    name: MIT
    url: https://github.com/goadesign/goa/blob/master/LICENSE
  title: The virtual wine cellar
  version: ""
paths:
  /cellar/accounts:
    get:
      description: Retrieve all accounts.
      operationId: account#list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/AccountCollection'
        "404":
          description: Not Found
      schemes:
      - http
      summary: list
    post:
      description: Create new account
      operationId: account#create
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CreateAccountPayload'
      responses:
        "201":
          description: Resource created
          headers:
            Location:
              description: href to created resource
              pattern: /accounts/[0-9]+
              type: string
      schemes:
      - http
      summary: create
  /cellar/accounts/{accountID}:
    delete:
      operationId: account#delete
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      schemes:
      - http
      summary: delete
    get:
      description: Retrieve account with given id. IDs 1 and 2 pre-exist in the system.
      operationId: account#show
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Account'
        "404":
          description: Not Found
      schemes:
      - http
      summary: show
    put:
      description: Change account name
      operationId: account#update
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/UpdateAccountPayload'
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      schemes:
      - http
      summary: update
  /cellar/accounts/{accountID}/bottles:
    get:
      description: List all bottles in account optionally filtering by year
      operationId: bottle#list
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      - description: Filter by years
        in: query
        items:
          type: integer
        name: years
        required: false
        type: array
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/BottleCollection'
        "404":
          description: Not Found
      schemes:
      - http
      summary: list
    post:
      description: Record new bottle
      operationId: bottle#create
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CreateBottlePayload'
      responses:
        "201":
          description: Resource created
          headers:
            Location:
              description: href to created resource
              pattern: ^/accounts/[0-9]+/bottles/[0-9]+$
              type: string
        "404":
          description: Not Found
      schemes:
      - http
      summary: create
  /cellar/accounts/{accountID}/bottles/{bottleID}:
    delete:
      operationId: bottle#delete
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      - in: path
        name: bottleID
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      schemes:
      - http
      summary: delete
    get:
      description: Retrieve bottle with given id
      operationId: bottle#show
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      - in: path
        name: bottleID
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Bottle'
        "404":
          description: Not Found
      schemes:
      - http
      summary: show
    patch:
      operationId: bottle#update
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      - in: path
        name: bottleID
        required: true
        type: integer
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/UpdateBottlePayload'
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      schemes:
      - http
      summary: update
  /cellar/accounts/{accountID}/bottles/{bottleID}/actions/rate:
    put:
      operationId: bottle#rate
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      - in: path
        name: bottleID
        required: true
        type: integer
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/RateBottlePayload'
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
      schemes:
      - http
      summary: rate
  /cellar/accounts/{accountID}/bottles/{bottleID}/watch:
    get:
      description: Retrieve bottle with given id
      operationId: bottle#watch
      parameters:
      - description: Account ID
        in: path
        name: accountID
        required: true
        type: integer
      - in: path
        name: bottleID
        required: true
        type: integer
      responses:
        "101":
          description: Switching Protocols
      schemes:
      - ws
      summary: watch
...
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
...
schemes:
- http
swagger: "2.0"